vk: TLAS creation, move init into a module
