+++ blade-graphics/src/util.rs
@@ -53,6 +53,7 @@ impl super::TextureFormat {
             Self::Rgba8Unorm => uncompressed(4),
             Self::Rgba8UnormSrgb => uncompressed(4),
             Self::Bgra8UnormSrgb => uncompressed(4),
+            Self::Rgba16Float => uncompressed(8),
             Self::Depth32Float => uncompressed(4),
         }
     }
